# docs/openapi.yml
openapi: 3.0.0

info:
  title: PickPin API
  version: 0.1.9
  contact:
    name: Grechko Georgy
    url: t.me/geogreck


servers:
  - url: http://localhost/api/v1
    description: Main server
  - url: http://localhost/api/v0
    description: Internal test server, powered by Express.js

tags:
  - name: Pins
  - name: Users
  - name: Profile
  - name: Posts
  - name: Auth
  - name: Boards
  - name: BoardPins

paths:
  /users/{user_id}:
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
    get:
      summary: Returns user with specified id
      responses:
        "200": # status code
          description: A JSON of user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      tags:
        - Users
      security:
        - cookieAuth: [ ]

  /users/{id}/profile:
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
    get:
      tags:
        - Profile
      summary: Get profile by user id
      description: Get profile by user id
      responses:
        "200": # status code
          description: A JSON of profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FullUpdateProfile"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "404":
          $ref: "#/components/responses/ErrResponseNoSuchUser"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]

  /profile:
    put:
      tags:
        - Profile
      summary: Full update of profile
      description: Full update of profile
      requestBody:
        description: Multipart form with values and avatar file
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FullUpdateProfile"
      responses:
        "200":
          description: A JSON object of updated profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
    patch:
      tags:
        - Profile
      summary: Partial update of profile
      description: Partial update of profile
      requestBody:
        description: Multipart form with values and avatar file
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PartialUpdateProfile"
      responses:
        "200":
          description: A JSON object of updated board.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]

  /posts:
    parameters:
      - schema:
          type: integer
        name: page
        in: query
        required: false
      - schema:
          type: integer
        name: limit
        in: query
        required: false
    get:
      summary: Returns a segment of posts according to the page and the number of pins in one page (limit parameter)
      responses:
        "200": # status code
          description: A JSON array of pins
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pin"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
      tags:
        - Posts
      security:
        - cookieAuth: [ ]

  /boards:
    post:
      tags:
        - Boards
      summary: Create a new board
      description: Create a new board
      requestBody:
        description: A JSON object containing the data to create board.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBoard"
      responses:
        "200":
          description: A JSON object of created board.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
      security:
        - cookieAuth: [ ]
    get:
      tags:
        - Boards
      summary: Returns all boards that the user has access to
      description: Returns all boards that the user has access to
      responses:
        "200":
          description: A JSON array of boards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boards"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
      security:
        - cookieAuth: [ ]

  /boards/{board_id}:
    get:
      tags:
        - Boards
      summary: Returns board with board_id
      description: Returns board with board_id
      responses:
        "200":
          description: A JSON object of board.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
      security:
        - cookieAuth: [ ]
    put:
      tags:
        - Boards
      summary: Full update of board
      description: Full update of board
      requestBody:
        description: A JSON object containing the data to update board.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FullUpdateBoard"
      responses:
        "200":
          description: A JSON object of updated board.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "403":
          $ref: "#/components/responses/ErrForbidden"
      security:
        - cookieAuth: [ ]
    patch:
      tags:
        - Boards
      summary: Partial update of board
      description: Partial update of board
      requestBody:
        description: A JSON object containing the data to update board.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartialUpdateBoard"
      responses:
        "200":
          description: A JSON object of updated board.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "403":
          $ref: "#/components/responses/ErrForbidden"
      security:
        - cookieAuth: [ ]
    delete:
      tags:
        - Boards
      summary: Delete board
      description: Delete board
      responses:
        "204":
          description: Board deleted successfully.
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "403":
          $ref: "#/components/responses/ErrForbidden"
      security:
        - cookieAuth: [ ]

  /pins:
    post:
      tags:
        - Pins
      summary: Create a new pin
      description: Create a new pin
      requestBody:
        description: Multipart form with values and image file
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreatePin"
      responses:
        "200":
          description: A JSON object of created pin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pin"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
    get:
      tags:
        - Pins
      summary: Returns all pins that the user has access to
      description: Returns all pins that the user has access to
      responses:
        "200":
          description: A JSON with pins array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pins"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]

  /users/{user_id}/pins:
    parameters:
      - schema:
          type: integer
        name: user_id
        in: path
        required: true
    get:
      tags:
        - Pins
      summary: Returns all pins of user
      description: Returns all pins of user
      responses:
        "200":
          description: A JSON with pins array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pins"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]

  /pins/{id}:
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
    get:
      tags:
        - Pins
      summary: Returns pin by id
      description: Returns pin by id
      responses:
        "200":
          description: A JSON object of pin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pin"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
    put:
      tags:
        - Pins
      summary: Full update of pin
      description: Full update of pin
      requestBody:
        description: Multipart form with values and image file
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FullUpdatePin"
      responses:
        "200":
          description: A JSON object of updated board.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pin"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "403":
          $ref: "#/components/responses/ErrForbidden"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
    delete:
      tags:
        - Pins
      summary: Delete pin
      description: Delete pin
      responses:
        "200":
          description: Pin deleted successfully.
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "403":
          $ref: "#/components/responses/ErrForbidden"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]

  /boards/{board_id}/pins:
    parameters:
      - schema:
          type: integer
        name: board_id
        in: path
        required: true
    get:
      tags:
        - BoardPins
      summary: Returns all pins of board
      description: Returns all pins of board
      responses:
        "200":
          description: A JSON with pins array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pins"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]

  /boards/{board_id}/pins/{id}:
    parameters:
      - schema:
          type: integer
        name: board_id
        in: path
        required: true
      - schema:
          type: integer
        name: id
        in: path
        required: true
    post:
      tags:
        - BoardPins
      summary: Add pin with "id" to board with "board_id"
      description: Add pin with "id" to board with "board_id"
      responses:
        "200":
          description: Pin added to board successfully.
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
    delete:
      tags:
        - BoardPins
      summary: Delete pin with "id" from board with "board_id"
      description: Add pin with "id" from board with "board_id"
      responses:
        "200":
          description: Pin deleted from board successfully.
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]

  /auth/signup:
    post:
      summary: Creates new user and returnes authentication cookie.
      security: [ ]
      tags:
        - Auth
      requestBody:
        required: true
        description: A JSON object containing the login and password.
        content:
          application/json:
            schema:
              allOf:
                - properties:
                    email:
                      $ref: "#/components/schemas/Email"
                - properties:
                    password:
                      $ref: "#/components/schemas/Password"
      responses:
        "204":
          description: >
            Successfully created user.
            Log in to get SESSION cookie. 
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"

  /auth/login:
    post:
      summary: Logs in and returns the authentication cookie
      requestBody:
        required: true
        description: A JSON object containing the login and password.
        content:
          application/json:
            schema:
              allOf:
                - properties:
                    email:
                      $ref: "#/components/schemas/Email"
                - properties:
                    password:
                      $ref: "#/components/schemas/Password"

      security: [ ]
      tags:
        - Auth
      responses:
        "200":
          description: >
            Successfully authenticated.
            The session ID is returned in a cookie named `JSESSIONID`. 
            You need to include this cookie in subsequent requests.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: JSESSIONID=abcde12345; Path=/; HttpOnly
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "404":
          $ref: "#/components/responses/ErrResponseNoSuchUser"
  /auth/logout:
    delete:
      summary: Logs out and deletes the authentication cookie.
      security:
        - cookieAuth: [ ]
      tags:
        - Auth
      responses:
        "204":
          description: Successfully logged out.
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "404":
          $ref: "#/components/responses/ErrResponseNoSuchCookie"
  /auth:
    post:
      summary: Updates existing authentication cookie.
      security:
        - cookieAuth: [ ]
      tags:
        - Auth
      responses:
        "200":
          description: >
            Successfully updated cookie.
            You need to include this cookie **now** in subsequent requests.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: JSESSIONID=abcde12345; Path=/; HttpOnly
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "404":
          $ref: "#/components/responses/ErrResponseNoSuchCookie"

components:
  schemas:
    Email:
      type: string
      format: email
      pattern: '/(?:[a-z0-9!#$%&*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&*+/=?^_`{|}~-]+)*|"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])/'
      example: vk@mail.ru
    Password:
      type: string
      format: password
      minLength: 8
      maxLength: 100
    Username:
      type: string
      minLength: 4
      maxLength: 40
      example: geogreck
    Name:
      type: string
      minLength: 1
      maxLength: 60
      example: Geogre
    Link:
      type: string
      minLength: 4
      maxLength: 100
      example: https://pickpin.ru
    MediaSource:
      type: string
      example: https://s3.pickpin.ru/uploaded_picture.jpg
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          $ref: "#/components/schemas/Email"
        username:
          $ref: "#/components/schemas/Username"
    Profile:
      type: object
      properties:
        username:
          $ref: "#/components/schemas/Username"
        name:
          $ref: "#/components/schemas/Name"
        profile_image:
          type: string
        website_url:
          type: string
    Board:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        privacy:
          type: string
          enum:
            - secret
            - public
          example: secret
        user_id:
          type: integer
      example: {
        "id": 3,
        "name": "Notes",
        "description": "Board for notes",
        "privacy": "secret",
        "user_id": 4
      }
    CreateBoard:
      type: object
      properties:
        name:
          type: string
          required: true
        description:
          type: string
          required: false
        privacy:
          type: string
          enum:
            - secret
            - public
          required: false
      example: {
        "name": "Notes",
        "description": "Board for notes",
        "privacy": "secret",
      }
    FullUpdateBoard:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        privacy:
          type: string
          enum:
            - secret
            - public
          example: secret
      example: {
        "name": "Notes",
        "description": "Board for notes",
        "privacy": "secret",
      }
    PartialUpdateBoard:
      type: object
      properties:
        name:
          type: string
          required: false
        description:
          type: string
          required: false
        privacy:
          type: string
          enum:
            - secret
            - public
          example: secret
          required: false
      example: {
        "name": "Notes"
      }
    Boards:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Board"
    FullUpdateProfile:
      type: object
      required:
        - username
        - name
        - website_url
        - bytes
      properties:
        username:
          $ref: "#/components/schemas/Username"
        name:
          $ref: "#/components/schemas/Name"
        website_url:
          type: string
        bytes:
          type: string
          format: binary
    PartialUpdateProfile:
      type: object
      properties:
        username:
          $ref: "#/components/schemas/Username"
        name:
          $ref: "#/components/schemas/Name"
        website_url:
          type: string
        bytes:
          type: string
          format: binary
    Pin:
      type: object
      properties:
        id:
          type: integer
        link:
          $ref: "#/components/schemas/Link"
        title:
          type: string
        description:
          type: string
        media_source:
          $ref: "#/components/schemas/MediaSource"
        author_id:
          type: integer
    Pins:
      type: object
      properties:
        pins:
          type: array
          items:
            $ref: "#/components/schemas/Pin"
    CreatePin:
      type: object
      required:
        - title
        - description
        - bytes
      properties:
        title:
          type: string
          description: Title
        description:
          type: string
          description: Description
        bytes:
          type: string
          format: binary
          description: Image file
    FullUpdatePin:
      type: object
      properties:
        title:
          type: string
          description: Title
        description:
          type: string
          description: Description
        bytes:
          type: string
          format: binary
          description: Image file
    Error:
      type: object
      properties:
        title:
          type: string
          description: Error title
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Verbose error details
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SESSIONID
  responses:
    ErrResponseBadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    ErrResponseNoSuchUser:
      description: User not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    ErrResponseNoSuchCookie:
      description: Cookie not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    ErrUnauthorized:
      description: Not authorized
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    ErrForbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    ErrInternalServerError:
      description: Internal Server Error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"